Enabling evil mode in Emacs.
I tried the evil vim emulation in Emacs before, but did not like it for
some reason. But the varying ctrl key across Linux, MacOS, BSD, RDP and
VNC is starting to drive me crazy. I can only use Emacs with the MacOS
cmd key, or with vim bindings, there, thats it.